requirements.txt
gradio==4.44.0
groq==0.15.0
elevenlabs==1.50.3
python-dotenv==1.0.1
Pillow==11.1.0
requests==2.32.3
pydub==0.25.1
SpeechRecognition==3.13.0
pyaudio==0.2.14
gtts==2.5.1